+++ packages/table-storybook/src/components/widgets/Phenotypes/Allele.js
